1 release ycompiler alpha2. 2 new yconfig by new ycompiler. 3 prepare version v0.2.0

+++ src/utility/yshell.cpp
@@ -36,12 +36,12 @@ static int8_t __execute_impl_linux(const std::vector<YLIB_STD_STRING> & cmd, con
     YLIB_STD_CHAR ** parse_cmd_array = new YLIB_STD_CHAR * [cmd.size() + args.size() + 1];
     parse_cmd_array[cmd.size() + args.size()] = nullptr;    
 
-    for(int _i = 0; _i < cmd.size(); _i++){
+    for(size_t _i = 0; _i < cmd.size(); _i++){
 
         parse_cmd_array[_i] = const_cast<char *>(cmd[_i].c_str());
     }
 
-    for(int _i = 0; _i < args.size(); _i++){
+    for(size_t _i = 0; _i < args.size(); _i++){
 
         parse_cmd_array[_i + cmd.size()] = const_cast<char *>(args[_i].c_str());
     }
@@ -50,7 +50,7 @@ static int8_t __execute_impl_linux(const std::vector<YLIB_STD_STRING> & cmd, con
     YLIB_STD_CHAR ** parse_cmd_env_array = new YLIB_STD_CHAR * [env.size() + 1];
     parse_cmd_env_array[env.size()] = nullptr;    
 
-    for(int _i = 0; _i < env.size(); _i++){
+    for(size_t _i = 0; _i < env.size(); _i++){
 
         parse_cmd_env_array[_i] = const_cast<char *>(env[_i].c_str());
     }
